backend/requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.1

# AI/ML and NLP
openai==1.3.7
langchain==0.0.340
spacy==3.7.2
scikit-learn==1.3.2
numpy==1.26.2
pandas==2.1.3
sentence-transformers==2.2.2

# Document parsing
pypdf2==3.0.1
python-docx==1.1.0
pdfplumber==0.10.3

# YouTube API
google-api-python-client==2.108.0

# Validation and utilities
pydantic==2.5.2
pydantic-settings==2.1.0
python-dotenv==1.0.0
httpx==0.25.2
aiofiles==23.2.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Redis for caching
redis==5.0.1
